Support "didMount" and "saveHook" on Renderable output
The name "didMount" was inspired by React. It's for running script on
the rendered element after it's in the DOM.

The "saveHook" is for deciding what gets saved to the worksheet. In
this new world, the output viewer has some of its own state that it
wants to save to the worksheet. If the output is interactive and the
user manipulates it, it's good to save these changes to the worksheet
somehow. Also this is useful for export, e.g. if the interactive
output is connected to a live server that won't be available when
someone else views this saved worksheet. In cases like this it's
useful to be able to export a canvas to PNG before save.