Remove margin when window is small. Enable side-scrolling in segments.
I use 980px as the max-width, and only use margins once we go beyond
that width. I chose that number because it's iOS's default screen
width, so it seems like a good target. Now you can render charts and
graphs and not worry about whether it'll overflow on an iPhone.