Multiple perf changes to be more compatible with phones
On save, convert Vega renderables to regular HTML using inline
SVG. This is much larger, but generating the SVG can destroy
phones. Instead, this change makes the file size large, and separately
I run a script pull out inline SVGs and PNGs and put them in files,
making the whole experience better.

To work alongside this, check for "foo.faster.clj" when given
"foo.clj", and serve it instead when it's available. The original path
is still needed because if the user wants to download the worksheet,
it should be self-containing.

So now I serve small worksheets which link to larger images. So
basically it's a normal webpage.

Remove the :value from all renderables. It doubles the size of
worksheets. This change is overkill, as it disables the alt+click
feature. But the alt+click feature doesn't work in the viewer
anyway. The real dissonance here is that there's a single "save"
that's trying to satisfy everybody. Exporting for the web has
different design constraints.